refactor(proxy): extract multi-range read handlers
Move MRR and read_range handling into multi_range_read.cc.
Document the custom batched primary-key MRR path on the handler declarations.
